Refactoring for BLS support, update unit tests , etc disable a couple of unit tests for now
