+++ src/catch_tests/denialofservice_tests.cpp
@@ -349,7 +349,7 @@ TEST_CASE("DoS_mapOrphans") {
     tx.vin[0].scriptSig << OP_1;
     tx.vout.resize(1);
     tx.vout[0].nValue = 1 * CENT;
-    tx.vout[0].scriptPubKey = GetScriptForDestination(key.GetPubKey().GetID());
+    tx.vout[0].scriptPubKey = GetScriptForDestination(key.GetPubKey().GetKeyID());
 
     AddOrphanTx(MakeTransactionRef(tx), i);
   }
@@ -363,7 +363,7 @@ TEST_CASE("DoS_mapOrphans") {
     tx.vin[0].prevout = COutPoint(txPrev->GetId(), 0);
     tx.vout.resize(1);
     tx.vout[0].nValue = 1 * CENT;
-    tx.vout[0].scriptPubKey = GetScriptForDestination(key.GetPubKey().GetID());
+    tx.vout[0].scriptPubKey = GetScriptForDestination(key.GetPubKey().GetKeyID());
     SignSignature(keystore, *txPrev, tx, 0, SigHashType());
 
     AddOrphanTx(MakeTransactionRef(tx), i);
@@ -376,7 +376,7 @@ TEST_CASE("DoS_mapOrphans") {
     CMutableTransaction tx;
     tx.vout.resize(1);
     tx.vout[0].nValue = 1 * CENT;
-    tx.vout[0].scriptPubKey = GetScriptForDestination(key.GetPubKey().GetID());
+    tx.vout[0].scriptPubKey = GetScriptForDestination(key.GetPubKey().GetKeyID());
     tx.vin.resize(2777);
     for (size_t j = 0; j < tx.vin.size(); j++) {
       tx.vin[j].prevout = COutPoint(txPrev->GetId(), j);
