+++ src/catch_tests/scriptflags.cpp
@@ -33,7 +33,7 @@ static std::map<std::string, uint32_t> mapFlagNames = {
     {"SIGHASH_FORKID", SCRIPT_ENABLE_SIGHASH_FORKID},
     {"REPLAY_PROTECTION", SCRIPT_ENABLE_REPLAY_PROTECTION},
     {"CHECKDATASIG", SCRIPT_VERIFY_CHECKDATASIG_SIGOPS},
-    {"SCHNORR", SCRIPT_ENABLE_SCHNORR},
+    //    {"BLS", SCRIPT_ENABLE_BLS},
 };
 
 uint32_t ParseScriptFlags(std::string strFlags) {
