+++ src/chainparams.cpp
@@ -99,7 +99,7 @@ public:
             "000000000000000003aadeae9dee37b8cb4838a866dae19b54854a0f039b03e0");
 
         // Date TBD in a future release
-        consensus.greatWallActivationTime = 1999999999;
+        consensus.blsActivationTime = 1999999999;
 
         /**
          * The message start string is designed to be unlikely to occur in
@@ -144,6 +144,7 @@ public:
         nExtCoinType = 339;
         cashaddrPrefix = "devault";
         cashaddrSecretPrefix = "dvtpriv";
+        blsAddrPrefix = "dvt";
 
         // Rewards
         consensus.nPerCentPerYear = {15,12,9,7,5};
@@ -201,7 +202,7 @@ public:
             "000000000000030bee568d677b6b99ee7d2d00b25d1fe95df5e73b484f00c322");
 
         // Date TBD in a future release
-        consensus.greatWallActivationTime = 1999999999;
+        consensus.blsActivationTime = 0;
 
         diskMagic[0] = 0x0d;
         diskMagic[1] = 0x08;
@@ -231,6 +232,7 @@ public:
         nExtCoinType = 1;
         cashaddrPrefix = "dvtest";
         cashaddrSecretPrefix = "testpriv"; // Shouldn't matter that "dv" isn't indicated
+        blsAddrPrefix = "blstest";
 
         // Rewards
         consensus.nPerCentPerYear = {1500,1200,900,7,5};
@@ -280,7 +282,7 @@ public:
         consensus.defaultAssumeValid = uint256S("0x00");
 
         // TBD
-        consensus.greatWallActivationTime = 1999999999;
+        consensus.blsActivationTime = 1999999999;
 
         diskMagic[0] = 0xfa;
         diskMagic[1] = 0xbf;
@@ -319,6 +321,7 @@ public:
         nExtCoinType = 1;
         cashaddrPrefix = "dvreg";
         cashaddrSecretPrefix = "regpriv";
+        blsAddrPrefix = "blsreg";
 
         // Rewards
         consensus.nPerCentPerYear = {15,12,9,7,5};
