+++ src/chainparams.h
@@ -60,6 +60,7 @@ public:
     int ExtCoinType() const { return nExtCoinType; }
     const std::string &CashAddrPrefix() const { return cashaddrPrefix; }
     const std::string &CashAddrSecretPrefix() const { return cashaddrSecretPrefix; }
+    const std::string &BLSAddrPrefix() const { return blsAddrPrefix; }
     const CCheckpointData &Checkpoints() const { return checkpointData; }
     const ChainTxData &TxData() const { return chainTxData; }
 
@@ -75,6 +76,7 @@ protected:
     int nExtCoinType;
     std::string cashaddrPrefix;
     std::string cashaddrSecretPrefix;
+    std::string blsAddrPrefix;
     std::string strNetworkID;
     CBlock genesis;
     bool fDefaultConsistencyChecks;
