+++ src/key.cpp
@@ -222,59 +222,6 @@ bool CKey::Derive(CKey &keyChild, ChainCode &ccChild, unsigned int nChild,
     return ret;
 }
 
-bool CExtKey::Derive(CExtKey &out, unsigned int _nChild) const {
-    out.nDepth = nDepth + 1;
-    CKeyID id = key.GetPubKey().GetKeyID();
-    memcpy(&out.vchFingerprint[0], &id, 4);
-    out.nChild = _nChild;
-    return key.Derive(out.key, out.chaincode, _nChild, chaincode);
-}
-
-void CExtKey::SetMaster(const uint8_t *seed, unsigned int nSeedLen) {
-    static const uint8_t hashkey[] = {'B', 'i', 't', 'c', 'o', 'i',
-                                      'n', ' ', 's', 'e', 'e', 'd'};
-    std::vector<uint8_t, secure_allocator<uint8_t>> vout(64);
-    CHMAC_SHA512(hashkey, sizeof(hashkey))
-        .Write(seed, nSeedLen)
-        .Finalize(vout.data());
-    key.Set(vout.data(), vout.data() + 32);
-    memcpy(chaincode.begin(), vout.data() + 32, 32);
-    nDepth = 0;
-    nChild = 0;
-    memset(vchFingerprint, 0, sizeof(vchFingerprint));
-}
-
-CExtPubKey CExtKey::Neuter() const {
-    CExtPubKey ret;
-    ret.nDepth = nDepth;
-    memcpy(&ret.vchFingerprint[0], &vchFingerprint[0], 4);
-    ret.nChild = nChild;
-    ret.pubkey = key.GetPubKey();
-    ret.chaincode = chaincode;
-    return ret;
-}
-
-void CExtKey::Encode(uint8_t code[BIP32_EXTKEY_SIZE]) const {
-    code[0] = nDepth;
-    memcpy(code + 1, vchFingerprint, 4);
-    code[5] = (nChild >> 24) & 0xFF;
-    code[6] = (nChild >> 16) & 0xFF;
-    code[7] = (nChild >> 8) & 0xFF;
-    code[8] = (nChild >> 0) & 0xFF;
-    memcpy(code + 9, chaincode.begin(), 32);
-    code[41] = 0;
-    assert(key.size() == 32);
-    memcpy(code + 42, key.begin(), 32);
-}
-
-void CExtKey::Decode(const uint8_t code[BIP32_EXTKEY_SIZE]) {
-    nDepth = code[0];
-    memcpy(vchFingerprint, code + 1, 4);
-    nChild = (code[5] << 24) | (code[6] << 16) | (code[7] << 8) | code[8];
-    memcpy(chaincode.begin(), code + 9, 32);
-    key.Set(code + 42, code + BIP32_EXTKEY_SIZE);
-}
-
 bool ECC_InitSanityCheck() {
     CKey key;
     key.MakeNewKey();
