+++ src/key.h
@@ -126,45 +126,6 @@ public:
               bool fSkipCheck);
 };
 
-struct CExtKey {
-    uint8_t nDepth;
-    uint8_t vchFingerprint[4];
-    unsigned int nChild;
-    ChainCode chaincode;
-    CKey key;
-
-    friend bool operator==(const CExtKey &a, const CExtKey &b) {
-        return a.nDepth == b.nDepth &&
-               memcmp(&a.vchFingerprint[0], &b.vchFingerprint[0],
-                      sizeof(vchFingerprint)) == 0 &&
-               a.nChild == b.nChild && a.chaincode == b.chaincode &&
-               a.key == b.key;
-    }
-
-    void Encode(uint8_t code[BIP32_EXTKEY_SIZE]) const;
-    void Decode(const uint8_t code[BIP32_EXTKEY_SIZE]);
-    bool Derive(CExtKey &out, unsigned int nChild) const;
-    CExtPubKey Neuter() const;
-    void SetMaster(const uint8_t *seed, unsigned int nSeedLen);
-    template <typename Stream> void Serialize(Stream &s) const {
-        unsigned int len = BIP32_EXTKEY_SIZE;
-        ::WriteCompactSize(s, len);
-        uint8_t code[BIP32_EXTKEY_SIZE];
-        Encode(code);
-        s.write((const char *)&code[0], len);
-    }
-    template <typename Stream> void Unserialize(Stream &s) {
-        unsigned int len = ::ReadCompactSize(s);
-        if (len != BIP32_EXTKEY_SIZE) {
-            throw std::runtime_error("Invalid extended key size\n");
-        }
-
-        uint8_t code[BIP32_EXTKEY_SIZE];
-        s.read((char *)&code[0], len);
-        Decode(code);
-    }
-};
-
 /**
  * Initialize the elliptic curve support. May not be called twice without
  * calling ECC_Stop first.
