+++ src/rpc/mining.cpp
@@ -6,7 +6,7 @@
 #include <rpc/mining.h>
 #include <amount.h>
 #include <blockvalidity.h>
-#include <cashaddrenc.h> // GetAddrFromTxOut for getblocktemplate
+#include <dstencode.h> // GetAddrFromTxOut for getblocktemplate
 #include <chain.h>
 #include <chainparams.h>
 #include <config.h>
