+++ src/wallet/wallet.cpp
@@ -47,7 +47,7 @@
 #include <variant>
 #include <optional>
 
-//#include <cashaddrenc.h>
+//#include <dstencode.h>
 
 
 std::vector<CWalletRef> vpwallets;
